+++ CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Moved gorilla tracing to a separate package
 - Renamed `context` package to `fxcontext`
 - Renamed `daemon` package to `fxdaemon`
-- Renamed `database/sql` package to `(fx)database/fxsql`
+- Renamed `database/sql` package to `fxsql`
+- Renamed `debug` package to `fxdebug`
+- Renamed `debug/metrics/prometheus` package to `fxpromhttp`
 
 
 ## [0.23.0] - 2018-04-29